repository.go,util.go: decouple edit w/ locking functionality
